refactor: Move MinigameRoutine and delete MinigameTask
